move registry insertion code
